Restructured
File structure added.
