--- src/pages/AboutPage.jsx
@@ -1,12 +1,0 @@
-import Navbar from "../components/Navbar";
-import Footer from "../components/Footer";
-
-export default function AboutPage() {
-  return (
-    <div>
-      <Navbar />
-      <h2>About Page</h2>
-      <Footer />
-    </div>
-  );
-}
