+++ src/pages/AboutPage/AboutPage.jsx
@@ -1,0 +1,12 @@
+import Navbar from "../../components/Navbar";
+import Footer from "../../components/Footer";
+
+export default function AboutPage() {
+  return (
+    <div>
+      <Navbar />
+      <h2>About</h2>
+      <Footer />
+    </div>
+  );
+}
