--- src/pages/LandingPage.jsx
@@ -1,17 +1,0 @@
-import Footer from "../components/Footer";
-import Navbar from "../components/Navbar";
-import Hero from "../components/Hero";
-import InfoSection from "../components/InfoSection";
-
-const LandingPage = () => {
-  return (
-    <>
-      <Navbar />
-      <Hero />
-      <InfoSection />
-      <Footer />
-    </>
-  );
-};
-
-export default LandingPage;
