+++ src/pages/LandingPage/LandingPage.jsx
@@ -1,0 +1,17 @@
+import Footer from "../../components/Footer";
+import Navbar from "../../components/Navbar";
+import Hero from "./components/Hero";
+import InfoSection from "./components/InfoSection";
+
+const LandingPage = () => {
+  return (
+    <>
+      <Navbar />
+      <Hero />
+      <InfoSection />
+      <Footer />
+    </>
+  );
+};
+
+export default LandingPage;
